refactor: consolidate types by moving ProjectParticipant and CreateProjectFormData to a shared types file
