+++ src/presentation/widgets/projects/types/index.ts
@@ -1,0 +1,12 @@
+export interface ProjectParticipant {
+  id?: string;
+  username?: string;
+  email: string;
+  avatarUrl?: string;
+}
+
+export interface CreateProjectFormData {
+  name: string;
+  description: string;
+  participants: string[];
+}
